requirements.txt
biopython>=1.80
streamlit>=1.28.0
